+++ improving thoughts.txt
@@ -1,5 +1,5 @@
 # -Akash Macha  09 December 2018 (Sunday)  21:22  PM
-
+# -Akash Macha  11 December 2018 (Tuesday)  16:53  PM
 ______________________________________
 Key points:
 1.  In master we've used normal functions to accomplish our task
